Recovered from Eric's Foolish Folly Push
After accidentally pushing an old version of Optimizer, this is the version from July 5th. I have also saved offline copies on email as needed of the following files that I interacted with (8 of the 9 didn't change):

Optimizer, Hub, Link, netcontrols, device, neural_network, Parallel, picoAmp, and dummy.

With this, everything should be recovered.